# specs/openapi.yaml
openapi: 3.0.3
info:
  title: HPM Patient Service API
  version: 1.0.0
  description: |
    The **HPM Patient Service** provides RESTful endpoints for managing patient demographic
    data within the Ai Nexus Hospital Patient Management (HPM) platform.

    ## Features
    - Patient registration with auto-generated Patient ID (`P{YEAR}{NNN}` format)
    - Paginated patient search with optional filters (status, gender, blood group, text search)
    - Patient profile retrieval with computed age
    - Patient demographic update
    - Patient lifecycle management (activate / deactivate)
    - Full audit trail via `X-User-ID` header

    ## Authentication
    This service does not implement authentication internally. All requests must pass
    through the Ai Nexus API Gateway which validates JWT tokens and forwards the
    verified user identity via the `X-User-ID` header.

    ## PHI Notice
    All patient data returned by this API is Protected Health Information (PHI) under HIPAA.
    Clients must handle responses in accordance with applicable privacy regulations.
  contact:
    name: Ai Nexus Platform Engineering
    email: engineering@ainexus.com
    url: https://www.ainexus.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8081
    description: Local development server
  - url: https://api-dev.ainexus.com/hpm
    description: Development environment
  - url: https://api.ainexus.com/hpm
...
paths:
  /api/v1/patients:

    post:
      tags:
        - Patient Management
      summary: Register a new patient
      operationId: registerPatient
      description: |
        Registers a new patient in the system. A unique Patient ID is automatically
        generated in the format `P{YEAR}{NNN}` (e.g., `P2026001`).

        If the submitted phone number already exists for another patient, the registration
        succeeds but the response includes a `warning` field indicating the duplicate.

        The `X-User-ID` header value is stored as `created_by` for the audit trail.
      parameters:
        - $ref: '#/components/parameters/XUserID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientRegistrationRequest'
            example:
              firstName: Jane
              lastName: Doe
              dateOfBirth: "1990-05-15"
              gender: FEMALE
              phone: "+1-555-867-5309"
              email: jane.doe@email.com
              bloodGroup: O_POS
              address: "123 Main Street, Apt 4B"
              city: Springfield
              state: IL
              zipCode: "62701"
              country: USA
              emergencyContactName: John Doe
              emergencyContactPhone: "+1-555-867-5310"
              emergencyContactRelationship: Spouse
              notes: Patient has latex allergy
      responses:
        '201':
          description: Patient registered successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PatientResponse'
              example:
                success: true
                message: Patient registered successfully
                warning: null
                data:
                  patientId: P2026001
                  firstName: Jane
                  lastName: Doe
                  dateOfBirth: "1990-05-15"
                  age: 35
                  gender: FEMALE
                  phone: "+1-555-867-5309"
                  email: jane.doe@email.com
                  bloodGroup: O_POS
                  status: ACTIVE
                  address: "123 Main Street, Apt 4B"
                  city: Springfield
                  state: IL
                  zipCode: "62701"
                  country: USA
                  emergencyContactName: John Doe
                  emergencyContactPhone: "+1-555-867-5310"
                  emergencyContactRelationship: Spouse
                  notes: Patient has latex allergy
                  createdAt: "2026-02-20T14:30:00Z"
                  createdBy: staff-001
                  updatedAt: null
                  updatedBy: null
                  deactivatedAt: null
                  deactivatedBy: null
                  activatedAt: null
                  activatedBy: null
        '400':
          description: Validation error — one or more fields failed validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
              example:
                success: false
                message: Validation failed
                errors:
                  phone: "Phone number must match one of the accepted formats: +1-XXX-XXX-XXXX, (XXX) XXX-XXXX, or XXX-XXX-XXXX"
                  dateOfBirth: "Date of birth must be a past date"
                timestamp: "2026-02-20T14:30:00Z"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    get:
      tags:
        - Patient Management
      summary: Search and list patients
      operationId: searchPatients
      description: |
        Returns a paginated list of patients. All filter parameters are optional and composable.
        Any combination of filters is valid, including no filters (returns all patients).

        The `search` parameter performs a case-insensitive partial match across:
        first name, last name, phone number, and patient ID.

        Results default to sorting by registration date descending (most recent first).
      parameters:
        - name: search
          in: query
          required: false
          description: Case-insensitive partial text search across first name, last name, phone, and patient ID
          schema:
            type: string
            example: "jane"
        - name: status
          in: query
          required: false
          description: Filter by patient status
          schema:
            $ref: '#/components/schemas/PatientStatus'
        - name: gender
          in: query
          required: false
          description: Filter by patient gender
          schema:
            $ref: '#/components/schemas/Gender'
        - name: bloodGroup
          in: query
          required: false
          description: Filter by blood group
          schema:
            $ref: '#/components/schemas/BloodGroup'
        - name: page
          in: query
          required: false
          description: Zero-indexed page number
          schema:
            type: integer
            minimum: 0
            default: 0
            example: 0
        - name: size
          in: query
          required: false
          description: Number of records per page (maximum 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
            example: 20
        - name: sort
          in: query
          required: false
          description: Sort field and direction (field,direction)
          schema:
            type: string
            default: "createdAt,desc"
            example: "lastName,asc"
      responses:
        '200':
          description: Patient search results
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PagedPatientSummaryResponse'
              example:
                success: true
                message: Patients retrieved successfully
                data:
                  content:
                    - patientId: P2026001
                      firstName: Jane
                      lastName: Doe
                      age: 35
                      gender: FEMALE
                      phone: "+1-555-867-5309"
                      bloodGroup: O_POS
                      status: ACTIVE
                    - patientId: P2026002
                      firstName: John
                      lastName: Smith
                      age: 42
                      gender: MALE
                      phone: "(555) 234-5678"
                      bloodGroup: A_POS
                      status: ACTIVE
                  pageable:
                    pageNumber: 0
                    pageSize: 20
                    sort:
                      sorted: true
                      direction: DESC
                      property: createdAt
                  totalElements: 2
                  totalPages: 1
                  last: true
                  first: true
                  numberOfElements: 2
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/patients/{patientId}:

    get:
      tags:
        - Patient Management
      summary: Get patient profile by ID
      operationId: getPatientById
      description: |
        Retrieves the complete patient profile for the given Patient ID.
        Returns all demographic, contact, clinical, and audit fields.
        The `age` field is computed from `dateOfBirth` at response time and is not stored.
      parameters:
        - $ref: '#/components/parameters/PatientIdPath'
      responses:
        '200':
          description: Patient profile retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PatientResponse'
              example:
                success: true
                message: Patient retrieved successfully
                data:
                  patientId: P2026001
                  firstName: Jane
                  lastName: Doe
                  dateOfBirth: "1990-05-15"
                  age: 35
                  gender: FEMALE
                  phone: "+1-555-867-5309"
                  email: jane.doe@email.com
                  bloodGroup: O_POS
                  status: ACTIVE
                  address: "123 Main Street, Apt 4B"
                  city: Springfield
                  state: IL
                  zipCode: "62701"
                  country: USA
                  emergencyContactName: John Doe
                  emergencyContactPhone: "+1-555-867-5310"
                  emergencyContactRelationship: Spouse
                  notes: Patient has latex allergy
                  createdAt: "2026-02-20T14:30:00Z"
                  createdBy: staff-001
                  updatedAt: "2026-03-15T09:00:00Z"
                  updatedBy: staff-002
                  deactivatedAt: null
                  deactivatedBy: null
                  activatedAt: null
                  activatedBy: null
        '404':
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "Patient not found with ID: P2026999"
                timestamp: "2026-02-20T14:30:00Z"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Patient Management
      summary: Update patient demographics
      operationId: updatePatient
      description: |
        Updates the mutable demographic fields of an existing patient.
        All fields in the request body replace the corresponding fields in the record.
        The `X-User-ID` header value is stored as `updated_by` and `updated_at` is set to now.

        Read-only fields (`patientId`, `createdAt`, `createdBy`, `status`) cannot be
        modified via this endpoint.
      parameters:
        - $ref: '#/components/parameters/PatientIdPath'
        - $ref: '#/components/parameters/XUserID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientUpdateRequest'
            example:
              firstName: Jane
              lastName: Doe-Smith
              dateOfBirth: "1990-05-15"
              gender: FEMALE
              phone: "+1-555-867-5309"
              email: jane.doesmith@email.com
              bloodGroup: O_POS
              address: "456 Oak Avenue"
              city: Chicago
              state: IL
              zipCode: "60601"
              country: USA
              emergencyContactName: John Doe-Smith
              emergencyContactPhone: "+1-555-867-5310"
              emergencyContactRelationship: Spouse
              notes: Patient has latex allergy. Updated address after relocation.
      responses:
        '200':
          description: Patient updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/ApiResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/PatientResponse'
              example:
                success: true
                message: Patient updated successfully
                data:
                  patientId: P2026001
                  firstName: Jane
                  lastName: Doe-Smith
                  dateOfBirth: "1990-05-15"
                  age: 35
                  gender: FEMALE
                  phone: "+1-555-867-5309"
                  email: jane.doesmith@email.com
                  bloodGroup: O_POS
                  status: ACTIVE
                  address: "456 Oak Avenue"
                  city: Chicago
                  state: IL
                  zipCode: "60601"
                  country: USA
                  emergencyContactName: John Doe-Smith
                  emergencyContactPhone: "+1-555-867-5310"
                  emergencyContactRelationship: Spouse
                  notes: Patient has latex allergy. Updated address after relocation.
                  createdAt: "2026-02-20T14:30:00Z"
                  createdBy: staff-001
                  updatedAt: "2026-03-15T09:00:00Z"
                  updatedBy: staff-002
                  deactivatedAt: null
                  deactivatedBy: null
                  activatedAt: null
                  activatedBy: null
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '404':
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/patients/{patientId}/deactivate:

    patch:
      tags:
        - Patient Management
      summary: Deactivate a patient record
      operationId: deactivatePatient
      description: |
        Sets the patient's status to `INACTIVE`. Deactivated patients are excluded from
        default search results but can still be retrieved by ID or by explicitly filtering
        with `status=INACTIVE`.

        This operation is idempotent: deactivating an already-inactive patient returns a
        descriptive message without error.

        The `X-User-ID` header value is stored as `deactivated_by` and `deactivated_at`
        is set to the current timestamp.
      parameters:
        - $ref: '#/components/parameters/PatientIdPath'
        - $ref: '#/components/parameters/XUserID'
      responses:
        '200':
          description: Patient deactivated successfully (or was already inactive)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                deactivated:
                  summary: Patient successfully deactivated
                  value:
                    success: true
                    message: Patient P2026001 has been deactivated
                    data: null
                already_inactive:
                  summary: Patient was already inactive
                  value:
                    success: true
                    message: Patient P2026001 is already inactive
                    data: null
        '404':
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/patients/{patientId}/activate:

    patch:
      tags:
        - Patient Management
      summary: Activate a patient record
      operationId: activatePatient
      description: |
        Sets the patient's status to `ACTIVE`. Reactivates a previously deactivated patient,
        making them visible in default search results again.

        This operation is idempotent: activating an already-active patient returns a
        descriptive message without error.

        The `X-User-ID` header value is stored as `activated_by` and `activated_at`
        is set to the current timestamp.
      parameters:
        - $ref: '#/components/parameters/PatientIdPath'
        - $ref: '#/components/parameters/XUserID'
      responses:
        '200':
          description: Patient activated successfully (or was already active)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                activated:
                  summary: Patient successfully activated
                  value:
                    success: true
                    message: Patient P2026001 has been activated
                    data: null
                already_active:
                  summary: Patient was already active
                  value:
                    success: true
                    message: Patient P2026001 is already active
                    data: null
        '404':
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
components:
  parameters:
    PatientIdPath:
      name: patientId
      in: path
      required: true
      description: Unique patient identifier in the format `P{YEAR}{NNN}` (e.g., `P2026001`)
      schema:
        type: string
        pattern: '^P\d{7}$'
        example: P2026001
    XUserID:
      name: X-User-ID
      in: header
      required: true
      description: |
        Identifier of the authenticated staff member performing the write operation.
        This value is stored in the audit trail (`created_by`, `updated_by`, etc.).
        Must be provided by the API Gateway after JWT validation; clients must not
        set this header directly in production.
      schema:
        type: string
        minLength: 1
        maxLength: 255
        example: staff-001
  schemas:
    PatientRegistrationRequest:
      type: object
      required:
        - firstName
        - lastName
        - dateOfBirth
        - gender
        - phone
        - bloodGroup
      properties:
        firstName:
          type: string
          minLength: 1
          maxLength: 100
          description: Patient's legal first name
          example: Jane
        lastName:
          type: string
          minLength: 1
          maxLength: 100
          description: Patient's legal last (family) name
          example: Doe
        dateOfBirth:
          type: string
          format: date
          description: Patient's date of birth. Must be a past date (ISO 8601 format)
          example: "1990-05-15"
        gender:
          $ref: '#/components/schemas/Gender'
        phone:
          type: string
          maxLength: 20
          pattern: '^(\+1-\d{3}-\d{3}-\d{4}|\(\d{3}\) \d{3}-\d{4}|\d{3}-\d{3}-\d{4})$'
          description: |
            Primary contact phone number. Accepted formats:
            - `+1-XXX-XXX-XXXX` (e.g., +1-555-867-5309)
            - `(XXX) XXX-XXXX` (e.g., (555) 867-5309)
            - `XXX-XXX-XXXX` (e.g., 555-867-5309)
          example: "+1-555-867-5309"
        email:
          type: string
          format: email
          maxLength: 255
          nullable: true
          description: Patient's email address (optional)
          example: jane.doe@email.com
        bloodGroup:
          $ref: '#/components/schemas/BloodGroup'
        address:
          type: string
          maxLength: 500
          nullable: true
          description: Street address (optional)
          example: "123 Main Street, Apt 4B"
        city:
          type: string
          maxLength: 100
          nullable: true
          description: City of residence (optional)
          example: Springfield
        state:
          type: string
          maxLength: 100
          nullable: true
          description: State or province (optional)
          example: IL
        zipCode:
          type: string
          maxLength: 20
          nullable: true
          description: Postal/ZIP code (optional, stored as string)
          example: "62701"
        country:
          type: string
          maxLength: 100
          nullable: true
          description: Country of residence (optional)
          example: USA
        emergencyContactName:
          type: string
          maxLength: 200
          nullable: true
          description: Full name of emergency contact (optional)
          example: John Doe
        emergencyContactPhone:
          type: string
          maxLength: 20
          nullable: true
          pattern: '^(\+1-\d{3}-\d{3}-\d{4}|\(\d{3}\) \d{3}-\d{4}|\d{3}-\d{3}-\d{4})$'
          description: Emergency contact phone number (optional, same format as phone)
          example: "+1-555-867-5310"
        emergencyContactRelationship:
          type: string
          maxLength: 100
          nullable: true
          description: Relationship of emergency contact to patient (optional)
          example: Spouse
        notes:
          type: string
          maxLength: 2000
          nullable: true
          description: General clinical or administrative notes (optional, may contain PHI)
          example: Patient has latex allergy
...
    PatientResponse:
      type: object
      description: Complete patient profile response including all demographic, contact, clinical, and audit fields
      properties:
        patientId:
          type: string
          description: System-generated unique patient identifier
          example: P2026001
        firstName:
          type: string
          example: Jane
        lastName:
          type: string
          example: Doe
        dateOfBirth:
          type: string
          format: date
          example: "1990-05-15"
        age:
          type: integer
          description: Computed age in years (calculated from dateOfBirth at response time; not stored)
          example: 35
        gender:
          $ref: '#/components/schemas/Gender'
        phone:
          type: string
          example: "+1-555-867-5309"
        email:
          type: string
          nullable: true
          example: jane.doe@email.com
        bloodGroup:
          $ref: '#/components/schemas/BloodGroup'
        status:
          $ref: '#/components/schemas/PatientStatus'
        address:
          type: string
          nullable: true
          example: "123 Main Street, Apt 4B"
        city:
          type: string
          nullable: true
          example: Springfield
        state:
          type: string
          nullable: true
          example: IL
        zipCode:
          type: string
          nullable: true
          example: "62701"
        country:
          type: string
          nullable: true
          example: USA
        emergencyContactName:
          type: string
          nullable: true
          example: John Doe
        emergencyContactPhone:
          type: string
          nullable: true
          example: "+1-555-867-5310"
        emergencyContactRelationship:
          type: string
          nullable: true
          example: Spouse
        notes:
          type: string
          nullable: true
          example: Patient has latex allergy
        createdAt:
          type: string
          format: date-time
          description: Timestamp of patient registration (UTC)
          example: "2026-02-20T14:30:00Z"
        createdBy:
          type: string
          description: User ID of staff member who registered the patient
          example: staff-001
        updatedAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of most recent demographic update (null if never updated)
          example: "2026-03-15T09:00:00Z"
        updatedBy:
          type: string
          nullable: true
          description: User ID of staff member who last updated demographics
          example: staff-002
        deactivatedAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of most recent deactivation (null if never deactivated)
          example: null
        deactivatedBy:
          type: string
          nullable: true
          description: User ID of staff member who deactivated the record
          example: null
        activatedAt:
          type: string
          format: date-time
          nullable: true
          description: Timestamp of most recent reactivation (null if never reactivated)
          example: null
        activatedBy:
          type: string
          nullable: true
          description: User ID of staff member who reactivated the record
          example: null
...
    ApiResponse:
      type: object
      description: Standard API response wrapper for all successful responses
      properties:
        success:
          type: boolean
          description: Indicates whether the operation succeeded
          example: true
        message:
          type: string
          description: Human-readable description of the operation result
          example: Patient registered successfully
        warning:
          type: string
          nullable: true
          description: Optional warning message (e.g., duplicate phone detected). Operation still succeeded.
          example: null
        data:
          description: The response payload (type depends on endpoint)
          nullable: true
    ValidationErrorResponse:
      type: object
      description: Response returned when Jakarta Bean Validation fails on the request body or parameters
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Validation failed
        errors:
          type: object
          description: Map of field names to validation error messages
          additionalProperties:
            type: string
          example:
            phone: "Phone number must match one of the accepted formats"
            dateOfBirth: "Date of birth must be a past date"
        timestamp:
          type: string
          format: date-time
          example: "2026-02-20T14:30:00Z"
    ErrorResponse:
      type: object
      description: Response returned for 4xx and 5xx errors
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Human-readable error description (does not expose internal implementation details or PHI)
          example: "Patient not found with ID: P2026999"
        timestamp:
          type: string
          format: date-time
          example: "2026-02-20T14:30:00Z"
    Gender:
      type: string
      enum:
        - MALE
        - FEMALE
        - OTHER
      description: Patient's gender identity
      example: FEMALE
    PatientStatus:
      type: string
      enum:
        - ACTIVE
        - INACTIVE
      description: Patient record lifecycle status
      example: ACTIVE
    BloodGroup:
      type: string
      enum:
        - A_POS
        - A_NEG
        - B_POS
        - B_NEG
        - AB_POS
        - AB_NEG
        - O_POS
        - O_NEG
        - UNKNOWN
      description: |
        ABO + Rh factor blood group classification.
        - `A_POS`: Blood type A, Rh positive
        - `A_NEG`: Blood type A, Rh negative
        - `B_POS`: Blood type B, Rh positive
        - `B_NEG`: Blood type B, Rh negative
        - `AB_POS`: Blood type AB, Rh positive (Universal Recipient)
        - `AB_NEG`: Blood type AB, Rh negative
        - `O_POS`: Blood type O, Rh positive
        - `O_NEG`: Blood type O, Rh negative (Universal Donor)
        - `UNKNOWN`: Blood group not determined or not provided
      example: O_POS